refactor: Move RealTimeQueries
